% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcfdr.R
\name{bcfdr}
\alias{bcfdr}
\title{Barber-Candès False Discovery Rate}
\usage{
bcfdr(p)
}
\arguments{
\item{p}{numeric vector of unadjusted p-values.}
}
\value{
Numeric vector of q-values of equal length and order
as the to input vector of p-values.
}
\description{
Simple implementation of the false discovery rate controlling procedure of
Barber and Candès (2015). The implementation is based on code available on
\href{https://github.com/lihualei71/AdaPT/blob/38a7d0a/R/other_methods.R}{GitHub}
by the authors of the \pkg{adaptMT} package as part of their analysis repository.

The function has been modified to return q-values for each p-value rather than
simply return the indices of rejected hypotheses at a pre-specified nominal FDR
threshold. The q-values are computed as the smallest nominal FDR threshold at
which each p-value would be rejected.
}
\examples{
## generate example p-values
pv <- runif(100)

## q-values using the BC procedure
qv <- bcfdr(p = pv)

}
\references{
References for the knockoff procedure for false discovery rate control (underlying theory):

Barber, R.F., Candès, E.J. (2015) Controlling the false discovery rate via knockoffs. The Annals of Statistics, 43(5):2055-2085. \url{https://doi.org/10.1214/15-AOS1337}
Arias-Castro, E., Chen, S. (2017) Distribution-free multiple testing. Electronic Journal of Statistics, 11(1):1983-2001. \url{https://doi.org/10.1214/17-EJS1277}

Reference for the original BC procedure function (underlying code):

Lihua, L. (2018) AdaPT paper repoistory. \url{https://github.com/lihualei71/AdaPT/}
}
\seealso{
\code{\link[knockoff]{knockoff}}
}
\author{
Patrick Kimes
}
